third_party/blink/web_tests/platform/generic/fast/forms/image/input-width-height-attributes-without-renderer-expected.txt

PASS inputElement.width is 50
PASS inputElement.height is 50
PASS successfullyParsed is true

TEST COMPLETE
